[Perf] More improvements to alias analysis
Summary:
Better handling of return:
 - kills all available modifications to local variables
 - maintains the aliasing state, rather than clearing it.

Test Plan:
fast_tests, slow_tests, flib/unit tests.

With inlining on, results in a 5% reduction in the number of inline-generated
variables in www-dev, a 0.5% reduction in the total number of Variant variables
declared, and a 1M reduction in the stripped binary size.

DiffCamp Revision: 150846
Reviewed By: qixin
CC: qixin, mwilliams, hphp-diffs@lists
Revert Plan:
OK