Open Source version still needs to optionally skip tcmalloc / jemalloc
Summary:
The open source version has problems with tcmalloc on Ubuntu 10.4 and jemalloc
on
FreeBSD and others.

Even though its a weak symbol we still require the headers to be included. For
now
it's easiest to just flip the defines so you have to defined code to *skip*
them which
we can add for the OSS version.

I'll look into bundling tcmalloc / jemalloc in the future when we just pick
one.

Test Plan:
make
make fast_tests

Check that with LD_PRELOAD that each can still be made available.

DiffCamp Revision: 155662
Reviewed By: ps
CC: ps, macvicar, hphp-diffs@lists
Revert Plan:
Ok