[Fix] Recreate ThreadSharedVariant if the SharedMap is updated
Summary:
When an array from APC contains an object, the object (and therefore the array)
might be changed without escalating the SharedMap. Therefore, whenever the
SharedVariant has the "should cache" flag, we need to recreate the
SharedVariant, rather than using the one that is wrapped by the SharedMap.

A testcase is added to capture the problem.

Test Plan:
fast_tests
TestCodeRun TestAPC

DiffCamp Revision: 168962
Reviewed By: qigao
CC: qigao, hphp-diffs@lists
Revert Plan:
OK