Change the meaning of the cached bit in StringData.
Summary:
We have a highest bit in StringData m_hash indicating whether it is
a "strictly integer" so that isStrictlyInteger can skip the call to
is_strictly_integer if a StaticString is NOT a "strictly integer".
From profiling it appears that is_numeric_string occurs more often
than is_strictly_integer(). I changed the meaning of the cached bit
from is_strictly_integer() to !is_numeric_string() with allow_errors == 1.
Finally, the cached bit is now extended to cover non-StaticString as long
as it is not a shared.

Test Plan:
make fast_tests
make slow_tests
make www and run perf-top

DiffCamp Revision: 170068
Reviewed By: qixin
CC: qixin, myang, hphp-diffs@lists
Revert Plan:
OK