Add PregenerateCPP option for clustering
The distribution of the code over the C++ files when clustering is
requested is often not very precise, because it performs the decisions
based on the size of the source PHP code, that might not be very
representative of the actual size that the output will have. The
subsequent repartition of the large files is helpful but still very
rough.

This patch adds an option to trade some additional memory use during
compilation for a more accurate partitioning. In case PregenerateCPP is
set to true, the output of all the files is pre-generated in memory,
and partitioning is directly done according to the size of the output.

PregenerateCPP is left to false by default.