Selectively decrease optimization level for long functions
This introduces some new options to decrease the compiler optimization
level for functions longer than a certain threshold, using the
attribute "optimize" that has been introduced in GCC 4.4. The patch
relies on the fact that CPP implementation files are generated before
the headers.

The options, GCCOptimization.O{2,1,0}, are disabled (set to 0) by
default.