[Fix] Fix order of eval for foreach
Summary:
Previous fix for evaluating $key and $value in the correct order, broke the
order of assignment.
The way hphpi currently works, theres no way to get both right. I assumed it
was more important to get the eval order correct, because nobody would write
foreach ($a as $x => $x) {}. But it turns out that they do.

This is just a bandaid so that if the key is a simple variable, we do the right
thing.

Test Plan:
fast_tests, flib/unit tests.

DiffCamp Revision: 171627
Reviewed By: iproctor
CC: iproctor, hphp-diffs@lists
Revert Plan:
OK