Add forward class header file
This adds a forward header file for classes. This file is intended to
declare all the constants, static strings, etc, that are required by
the class header.

This patch is a preparation for making the class header files
completely self-contained, so that it is not required to include
anything else before a class header file in order to get valid C++
code. This way the inclusion system can be simpler.