Fix race condition of fb_const_fetch
Summary:
The ZendArray const_data wasn't fully protected. We need to set both itself
and all of its contents to be static. Added call to onStatic() to ensure that.

Test Plan:
make fast_tests
use a apc_prime.so from a production box and verify that the const_data
indeed is set to static.

DiffCamp Revision: 174623
Reviewed By: qixin
CC: qixin, hphp-diffs@lists
Tasks:
424098: fb_const_fetch corrupt on some of make-haste keys.

Revert Plan:
OK