New way to define sandboxes
Summary:
An alternative way to map urls to directories.
Rather than looking for .hphp file in /home/<user> it defines a root
directory containing all the sandboxes with directory names matching the
subdomains.

Test Plan:
Tested that without the FromCommonRoot Sandbox option it behaves the
same as the current version.
With the flag set it loads the the source from the
DirectoriesRoot/<sandbox_name> directory,
and writes the error/access logs to files under LogsRoot.
If the directory doesn't exist it returns a "Not Found" page.

DiffCamp Revision: 178524
Reviewed By: hzhao
Commenters: pfung
CC: pfung, hzhao, hphp-diffs@lists, slawekbiel

Revert Plan:
OK