[RPC] Support include and include_once for files
Summary:
When we added the support for invoking a file over RPC, it was always using the
include_once semantics, which was fine for providing a definition for a
function, but not if we were really trying to invoke a file.

This diff changes the file inclusion parameter to "include=" and
"include_once=" to make it obvious whether the file is included once. Also, if
the RPC command is just a PHP file, it is equivalent to "include=".

Also updated the doc.

Test Plan:
Feng's testcase. All one hundred requests succeeded.

Also tested the "once" parameter.

DiffCamp Revision: 182173
Reviewers: hzhao
CC: hzhao, fengqian, hphp-diffs@lists
Tasks: #370148
Revert Plan:
OK

Blame Revision: