[Fix] Don't clean up the PREGCache at shutdown time
Summary:
The preg cache is now thread-local instead of request-local, and we shouldn't
do the cleanup in its destructor any more, because we don't have a way to
specify the order of destruction of thread-local objects, which means that the
cache might be destructed after all the StringData inside it have been deleted,
causing crashes at process shutdown time.

We should be able to just not do the cleanup, because the process is gone any
way.

Test Plan:
Built hphpi and put it on devrs109. No more coredumps reported.
Built www and put it on one h2e machine. No pcre coredump when the process was
shutdown.

DiffCamp Revision: 182233
Reviewed By: hzhao
CC: hzhao, qixin, hphp-diffs@lists
Revert Plan:
OK