[Fix] Type checked params in constructors didnt behave correctly.
Summary:
My recent fixes for type checked parameters fixed the constructor, but not the
create function which calls the constructor. The result was that arguments were
coerced to CArrRef, or p_Class before being passed to the constructor, which
could cause it to silently accept arguments, or reject them with the wrong
error.

Also fixed the error reporting to report 1 based parameter indices, and to
report the function name correctly.

Test Plan:
fast_tests, slow_tests, new test case. Verify correct error for the ajax url in
the task.

DiffCamp Revision: 183277
Reviewed By: qixin
CC: achao, qixin, hphp-diffs@lists
Tasks:
434404: HPHP allows null through class typehint

Revert Plan:
OK