[Fix] Update dependencies in AnalyzeFinal
Summary:
This makes it easier when optimizations cause new dependencies. In particular,
inlining was causing classes to be referenced, but the relevant headers were not
included.

Also cleanup the code. Passing "false" for the strong parameter caused
addUserFunction and addUserClass to be no-ops (due to other code cleanup). So
remove the param, and remove the calls which passed false.

Test Plan:
build www-dev, with and without --opts=all

DiffCamp Revision: 183358
Reviewed By: hzhao
CC: hzhao, hphp-diffs@lists
Revert Plan:
OK