[Fix] Do not create a ref wrapper for passed-in CVarRef in list assignment
Summary:
The ref wrapper is to handle the case where the variable being
assigned in list assignment overlaps with the array itself. If the array
is a parameter passed in as CVarRef, it cannot be such a case.

Test Plan:
compile www-dev and compare the diff in CPP

DiffCamp Revision: 183605
Reviewed By: mwilliams
CC: qixin, mwilliams, hphp-diffs@lists
Revert Plan:
OK