Fix link creation bug in cronolog
Summary:
This bug existed in the original cronolog source code. The problem is that
sometimes the target file has been aged out and removed, but the dangling
symlink is never updated because stat() returned -1 when the target file is
missing. As a result the link itself isn't removed and subsequent new link
could not be setup to point to the new log file.

Test Plan:
make fast_tests
throw the new binary onto a production box that had the problem, enable
native cronolog support and observe the dangling link get updated correctly.

DiffCamp Revision: 192427
Reviewed By: hzhao
Reviewers: hzhao
CC: hzhao, myang, hphp-diffs@lists
Revert Plan:
OK