[Compile-time] Remove code related to LastInference
Summary:
We do a full pass of type inference with AnalysisResult::m_phase set to
LastInference. This takes a significant part of the time, but does very little.

Some of the things it does turn out to be counter-productive, so remove them.
Most of the rest can be moved to AnalyzeFinal. Setting the return type of a
function turned out to be tricky... so just build a list of all the return
expressions in a function, and post process them.

Test Plan:
fast_tests, slow_tests, flib/unit tests. Inspect differences in www-dev
generated code.

This reduces the inferTypes time for www-dev from about 65s to about 40s, and
also fixes a number of small bugs in the dependencies which were largely masked
by always doing one extra pass over all the functions at the end of inferTypes.

DiffCamp Revision: 193437
Reviewed By: hzhao
Reviewers: hzhao
CC: hzhao, hphp-diffs@lists
Revert Plan:
OK