[Fix] Exit with broken pipe for cli
Summary:
PHP stops execution when SIGPIPE comes, at least for echo. This diff enables
the cli mode of hphp/hphpi to exit when there is broken pipe, controlled by
cli.ExitOnSigPipe, which is true by default. Therefore, piping scripts together
would not be hanging when the receiver end is gone.

PHP doesn't seem to exit with fwrite() when SIGPIPE comes, which seems to be a
random decision. We are not differentiating these cases here. I think it is
better not to hang anyway.

Test Plan:
Pipe a script to another one that throws. No more hanging.

DiffCamp Revision: 194042
Reviewed By: hzhao
Reviewers: hzhao
CC: hzhao, hphp-diffs@lists
Tasks:
455441: Issues with script under hphpi

Revert Plan:
OK