Include command-line arguments in the stacktrace
Summary:
When the webserver crashes, we have the URL in the stacktrace file, but when
the command-line crashes, we don't have the parameters. This diff adds the
parameters to the stacktrace file in cli mode.

Test Plan:
built hphpi, caused a crash, and looked at the stacktrace file

DiffCamp Revision: 196091
Reviewed By: hzhao
Reviewers: hzhao
CC: hzhao, hphp-diffs@lists
Revert Plan:
OK