Add function fb_urandom
Summary:
The file /dev/urandom is used to generate random bytes. If we read the file
too fast, the kernel has to recharge it. The faster the file is read the more
kernel CPU is spent. I wrote one test that keep reading /dev/urandom. When
running on a production machine I saw that the kernel indeed spent more than
13% CPU:
    11.17%            a.out  [kernel]
               [k] sha_transform
     1.68%            a.out  [kernel]
               [k] mix_pool_bytes_extract
     0.45%            a.out  [kernel]
               [k] extract_buf
In HPHP, dev/urandom is treated as a PlainFile and we always read 8k in one
shot. PHP does the same. If we change 8k to smaller number we can have
different behavior from PHP. However, in one request we may only need tens
of random bytes so the rest of the 8k bytes read would be wasted. I added a
new function fb_urandom which uses a persistent Object to store the opened
/dev/urandom file handle across requests. In this way we don't waste random
bytes so the kernel can spend less time for recharging /dev/urandom.

Test Plan:
make fast_tests
build www (updated to use the new function) and run on a API machine, run
perf-top and saw 0.00% of the above kernel functions.

DiffCamp Revision: 197376
Reviewed By: hzhao
Reviewers: hzhao, qixin
Commenters: ps, bmaurer
CC: bmaurer, ps, hzhao, myang, hphp-diffs@lists
Revert Plan:
OK